requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.2
typing-extensions==4.12.2
gunicorn==20.1.0
